requirements.txt
datasets>=2.10.1
torch>=1.13.0
transformers>=4.31.0
sentencepiece>=0.1.99
matplotlib>=3.8.0
numpy>=1.26.0
pandas>=2.1.1
scikit-learn>=1.3.1
protobuf>=4.23.4
wandb>=0.16.1
ipykernel==6.29.5
python-Levenshtein==0.12.0
accelerate>=0.30.1